React native: separated native stories to own files
