+++ .gitignore
@@ -6,3 +6,4 @@ storybook-static/
 .idea/
 build/
 lib/
+storybook/storyLoader.js
