+++ stories_native/Button.test.js
@@ -1,0 +1,5 @@
+import initStoryshots from '@storybook/addon-storyshots';
+
+initStoryshots({
+  storyKindRegex: /^Button$/
+});
