+++ stories_native/Newsfeed.test.js
@@ -1,3 +1,5 @@
 import initStoryshots from '@storybook/addon-storyshots';
 
-initStoryshots();
+initStoryshots({
+  storyKindRegex: /^NewsFeed$/
+});
